Refactoring: class GeoDataPointmap::AddPointSetReferenceCommand moved to independent files.
